\name{ttr.results}
\alias{ttr.results}
\title{ Technical Trading Rule Results }
\description{
  Summarize the results of a price series and signals.
}
\usage{
  ttr.results(beg, end, sig, price)
}
\arguments{
  \item{beg}{ Numeric; The beginning date, in CCYYMMDD format. }
  \item{end}{ Numeric; The ending date, in CCYYMMDD format. }
  \item{sig}{ A vector of signals; use \sQuote{0} for no position, \sQuote{1} for
    long positions, and \sQuote{-1} for short positions. }
  \item{price}{ A vector of prices to apply the signals to. }
}
\details{
  I'm sure there are details, but I haven't added them yet... ;)
}
\value{
  A table showing the results for the period \code{beg} to \code{end}, containing the columns:
  \tabular{ll}{
    Package: \tab TTR\cr
    Strategy: \tab Long Only, Long and Short, Long with Margin, Long and Short
      with Margin, and Buy and Hold.\cr
    Rule Return: \tab The _annualized_ TTR return for each strategy.\cr
    Rule-B&H: \tab The _annualized_ return for each strategy minus the Buy and Hold return.\cr
    Sharpe Ratio: \tab The Sharpe Ratio for each strategy.\cr
    StDev Ratio: \tab The Standard Deviation Ratio for each strategy.\cr
  }

  And a table showing other results with the following rows:
  \tabular{ll}{
    Trades/Year: \tab The average number of trades per year generated by the TTR.\cr
    Avg Lgth Buy: \tab The average length - in days - of the long trades generated by the TTR.\cr
    Avg Lgth Sell: \tab The average length - in days - of the short trades generated by the TTR.\cr
    Min/Max Buy: \tab The minimum / maximum length - in days - of the long trades
      generated by the TTR.\cr
    Min/Max Sell: \tab The minimum / maximum length - in days - of the short trades
      generated by the TTR.\cr
  }
}
\author{ Josh Ulrich }
\note{
  This function is likely to be changed in the future.
}
\examples{
  ## None at this time.
}
\keyword{ ts }